refactor(select): rename typings to types and run formatter
