+++ packages/select/.npmignore
@@ -1,0 +1,1 @@
+**/*test.js
