+++ packages/select/index.js
@@ -1,5 +1,3 @@
-import Select from './src/Select';
-
+export { default } from './src/Select';
 export { default as SelectField } from './src/SelectField';
 export { default as ResourceSelect } from './src/ResourceSelect';
-export default Select;
