+++ packages/select/package.json
@@ -3,7 +3,7 @@
   "version": "0.18.1",
   "description": "Wrapper for react-select to work with formik",
   "main": "index.js",
-  "typings": "typings/index.d.ts",
+  "types": "index.d.ts",
   "keywords": [
     "availity",
     "reactstrap",
@@ -24,8 +24,6 @@
     "@availity/api-axios": "^5.5.14",
     "@availity/api-core": "^7.0.2",
     "@availity/form": "^0.7.0",
-    "@availity/localstorage-core": "^3.0.0",
-    "@types/react-select": "^3.0.23",
     "axios": "^0.21.1",
     "formik": "^2.2.8",
     "react": "^17.0.2",
@@ -34,16 +32,16 @@
     "yup": "^0.30.0"
   },
   "peerDependencies": {
-    "@availity/api-axios": "^3.0.1",
-    "@availity/api-core": "^3.0.0",
-    "@availity/form": "^0.2.0 || ^0.3.0 || ^0.4.0  || ^0.5.0",
-    "@availity/localstorage-core": "^2.5.0",
-    "axios": "^0.19.2",
+    "@availity/api-axios": "^5.0.0",
+    "@availity/api-core": "^7.0.0",
+    "@availity/form": "^0.2.0 || ^0.3.0 || ^0.4.0  || ^0.5.0 || ^0.6.0 || ^0.7.0",
+    "axios": "^0.21.1",
     "formik": "^2.0.1",
     "react": ">=16.8.3",
     "reactstrap": "^8.0.0 || ^7.0.0 || ^6.0.0"
   },
   "dependencies": {
+    "@types/react-select": "^3.0.23",
     "classnames": "^2.3.1",
     "lodash": "^4.17.21",
     "prop-types": "^15.7.2",
